Added some primitive 3 address code generation
The code for FOR statements is known to be incorrect.

Still debating whether the code as implemented
is going into too much detail with the explicit address
calculation.  It's probably ok, and I'm just getting
cheesed off at how verbose it makes some of the
generation.  I may need to revisit if it
overcomplicates optimization of the intermediate
code.